sample_base message suffix req/res 0.0.64

+++ sinnori/project/sample_base/client_build/web_build/src/main/java/kr/pe/sinnori/servlet/BoardListSvl.java
@@ -28,9 +28,9 @@ public class BoardListSvl extends AbstractServlet {
 		if (null == parmBoardId) parmBoardId = "2";
 		
 		
-		long boardId = 2L;		
+		short boardId = 2;		
 		try {
-			boardId = Long.parseLong(parmBoardId);
+			boardId = Short.parseShort(parmBoardId);
 		}catch (NumberFormatException nfe) {
 			String errorMessage = new StringBuilder("자바 long 타입 변수인 게시판 식별자(boardId) 값[")
 			.append(parmBoardId).append("]이 잘못되었습니다.").toString();
