+++ sinnori/project/sample_base/client_build/web_build/src/main/java/kr/pe/sinnori/servlet/BoardWriteSvl.java
@@ -111,10 +111,10 @@ public class BoardWriteSvl extends AbstractLoginServlet {
 				return;
 			}
 			
-			long boardId = 2L;
+			short boardId = 2;
 
 			try {
-				boardId = Long.parseLong(parmBoardId);
+				boardId = Short.parseShort(parmBoardId);
 			}catch (NumberFormatException nfe) {
 				String errorMessage = new StringBuilder("자바 long 타입 변수인 게시판 식별자 값[")
 				.append(parmBoardId).append("]이 잘못되었습니다.").toString();
