+++ sinnori/project/sample_base/server_build/src/main/java/kr/pe/sinnori/server/lib/BoardType.java
@@ -1,17 +1,17 @@
 package kr.pe.sinnori.server.lib;
 
 public enum BoardType {
-	NOTICE((byte)0, "공지"), FREE((byte)1, "자유"), FAQ((byte)2, "질문과답변");
+	NOTICE((short)0, "공지"), FREE((short)1, "자유"), FAQ((short)2, "질문과답변");
 	
-	private byte boardTypeValue;
+	private short boardTypeValue;
 	private String boardTypeName;
 	
-	private BoardType(byte boardTypeValue, String boardTypeName) {
+	private BoardType(short boardTypeValue, String boardTypeName) {
 		this.boardTypeValue = boardTypeValue;
 		this.boardTypeName = boardTypeName;
 	}
 	
-	public byte getValue() {
+	public short getValue() {
 		return boardTypeValue;
 	}
 	
@@ -19,7 +19,7 @@ public enum BoardType {
 		return boardTypeName;
 	}
 	
-	public static BoardType valueOf(byte boardTypeValue) {
+	public static BoardType valueOf(short boardTypeValue) {
 		BoardType[] boradTypes = BoardType.values();
 		for (BoardType boardType : boradTypes) {
 			if (boardType.getValue() == boardTypeValue) {
